Fix encoding issues with passwords
FreeRADIUS handles passwords as byte-strings. How the client or the
plugin encode/decode these bytes is up to them.
Some VPNs seem to use ISO-latin encoding, other UTF-8.
With this fix we try to guess the encoding using a built-in perl
module and decode it accordingly before URL-encoding it for privacyIDEA.